requirements.txt
fastapi
uvicorn
jinja2
pydantic
python-dotenv
google-generativeai
textblob